requirements.txt
python-binance==1.0.17
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
streamlit==1.35.0
twilio==9.0.4
seaborn==0.13.2
apscheduler==3.10.4
cryptography==42.0.7
streamlit-autorefresh
